requirements.txt
Flask==3.1.1
flask_mail==0.10.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
psycopg2_binary==2.9.10
qrcode==8.2
Werkzeug==3.1.3
